sys/src/9/vt4: fix prototype of ainc and adec

+++ sys/src/9/vt5/fns.h
@@ -8,8 +8,8 @@ typedef struct Ether Ether;
 #pragma incomplete Dma
 
 void	addconf(char*, char*);
-int	adec(long*);
-int	ainc(long*);
+long	adec(long*);
+long	ainc(long*);
 void	archetherspeed(int, int);
 void	archinit(void);
 uint	archmiiport(int);
